modules/coreboot: update to version 4.12
Update coreboot to version 4.12 to support vboot measure boot.

Warning:
This commit does only apply the basic requirement for vboot targets.
It does not adapt any board configurations and also disables the use of
the crosstoolchain for coreboot. Therfore it shouldn't be merged with
the master branch at the current state.

Signed-off-by: Marcello Sylvester Bauer <sylv@sylv.io>